Modularization of the utilities
Considering that the utilities are involved even in the lookup process,
it makes sense to modularize them for usage.
